Refactored App component
