+++ movies-app/src/MovieListSection/index.js
@@ -1,0 +1,3 @@
+import MovieListSection from './MovieListSection';
+
+export default MovieListSection;
